+++ src/client/routes/posts/posts-list/post-card/post-card.component.tsx
@@ -1,11 +1,19 @@
 import Link from 'next/link';
-import type { FC } from 'react';
+import maybe, { Maybe } from 'fnts/maybe';
+import fold from 'fnts/maybe/operators/fold';
+import { FC, ReactNode, useMemo } from 'react';
 
+import { useToggle } from '@/shared/hooks';
 import type { ListPost } from 'shared/entity';
 import { Meta } from '@/routes/posts/shared/components';
 import { lazyLoad, tryVibrate } from '@/shared/utils';
 
 import * as Styled from './post-card.styles';
+
+const ImageSkeleton = lazyLoad(
+  'ImageSkeleton',
+  () => import('@/routes/posts/shared/components/image-skeleton')
+);
 
 const Markdown = lazyLoad(
   'Markdown',
@@ -22,6 +30,24 @@ export const PostCard: FC<PostCardProps> = ({
   date,
   readTime,
 }) => {
+  const [hasImageLoaded, toggleImage] = useToggle();
+
+  const maybeImage = useMemo<Maybe<ReactNode>>(() => {
+    return maybe(
+      imageURL && (
+        <>
+          <Styled.Hero
+            src={imageURL}
+            alt={title}
+            onLoad={() => toggleImage(true)}
+          />
+
+          {!hasImageLoaded && <ImageSkeleton />}
+        </>
+      )
+    );
+  }, [imageURL, title, toggleImage, hasImageLoaded]);
+
   return (
     <Styled.PostCard>
       <Link href={`/p/${id}`} passHref>
@@ -29,7 +55,7 @@ export const PostCard: FC<PostCardProps> = ({
           // eslint-disable-next-line functional/prefer-tacit
           onClick={() => tryVibrate()}
         >
-          {imageURL && <Styled.Hero src={imageURL} alt={title} />}
+          {fold(maybeImage)}
 
           <Styled.Title>{title}</Styled.Title>
         </Styled.Link>