+++ src/client/routes/posts/shared/components/index.ts
@@ -1,3 +1,5 @@
 export * from './list';
 export * from './meta';
 export * from './dot-divider';
+export * from './image-skeleton';
+export { Skeleton } from './skeleton';