+++ src/client/routes/posts/shared/components/skeleton/utils/parse-height.ts
@@ -7,17 +7,18 @@ export const parseHeight = (
 ): string | undefined => {
   switch (height) {
     case 'h1': {
-      return '2.5rem';
+      return '2.074rem';
     }
     case 'h2': {
-      return '2rem';
+      return '1.728rem';
     }
     case 'h3': {
-      return '1.5rem';
+      return '1.44rem';
     }
     case 'h4': {
-      return '1.1rem';
+      return '1.2rem';
     }
+    case 'h5':
     case 'p': {
       return '1rem';
     }